separate wiggle for gc and scores
